requirements.txt
discord.py==2.3.2
parsedatetime==2.6
pytest==7.4.3

celery~=5.3.6
requests~=2.31.0
